.codex/skills/cli-tool-skills/pm-99-utils-and-standards/requirements.txt
pm4py>=2.7.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
pyyaml>=6.0
